🐛 Bug
fix auth

+++ src/watchlist/watchlist.controller.ts
@@ -10,6 +10,7 @@ import {
   UnauthorizedException,
   Query,
   UseFilters,
+  Put,
 } from '@nestjs/common';
 import { WatchlistService } from './watchlist.service';
 import { AuthGuard } from 'src/auth/auth.guard';
@@ -20,7 +21,10 @@ import {
   ApiResponse,
   ApiQuery,
 } from '@nestjs/swagger';
-import { CreateWatchlistDto } from 'src/utils/Watchlist.dto';
+import {
+  CreateWatchlistDto,
+  UpdateWatchlistDto,
+} from 'src/utils/Watchlist.dto';
 import { PrismaExceptionFilter } from 'src/common/exceptions/prisma-exception.filter';
 
 @Controller('watchlist')
@@ -51,6 +55,28 @@ export class WatchlistController {
     return await this.watchlistService.createWatchList(user.sub, data);
   }
 
+  @Put()
+  @ApiOperation({ summary: 'Update watchlist' })
+  @ApiResponse({
+    status: 201,
+    description: 'Content successfully updated to watchlist.',
+  })
+  @ApiResponse({ status: 400, description: 'Invalid request parameters.' })
+  async update(
+    @Req() request: Request,
+    @Body() data: UpdateWatchlistDto,
+  ): Promise<any> {
+    const user = request['user'];
+
+    if (!user) {
+      throw new UnauthorizedException('User not found');
+    }
+
+    await this.watchlistService.updateWatchList(user.sub, data);
+
+    return 'Content successfully updated to watchlist.';
+  }
+
   @Get()
   @ApiOperation({ summary: 'Get all watchlist entries with pagination' })
   @ApiQuery({
@@ -69,39 +95,39 @@ export class WatchlistController {
   async findAll(
     @Query('page') page: string = '1',
     @Query('limit') limit: string = '10',
+    @Req() request: Request,
   ) {
-    return await this.watchlistService.getAllWatchList(
+    const user = request['user'];
+
+    if (!user) {
+      throw new UnauthorizedException('User not found');
+    }
+
+    return await this.watchlistService.getWatchlistByUser(
+      user.sub,
       Number(page),
       Number(limit),
     );
   }
 
-  @Get('user/:userId')
-  @ApiOperation({ summary: 'Get watchlist of a user with pagination' })
-  @ApiQuery({ name: 'page', required: false, example: 1, description: 'Page number' })
-  @ApiQuery({ name: 'limit', required: false, example: 10, description: 'Number of items per page' })
-  @ApiResponse({ status: 200, description: 'Successfully fetched user watchlist.' })
-  @ApiResponse({ status: 404, description: 'User watchlist not found.' })
-  async findByUser(
-    @Param('userId') userId: string,
-    @Query('page') page: string = '1',
-    @Query('limit') limit: string = '10',
-  ) {
-    return await this.watchlistService.getWatchlistByUser(Number(userId), Number(page), Number(limit));
-  }
-
-  @Get('entry/:id')
+  @Get('/:id')
   @ApiOperation({ summary: 'Get one watchlist by id' })
   @ApiResponse({
     status: 200,
     description: 'Successfully fetched watchlist.',
   })
-  @ApiResponse({ status: 404, description: 'watchlist not found.' })
-  async getWatchlistById(@Param('id') id: string) {
+  @ApiResponse({ status: 404, description: 'Watchlist not found.' })
+  async getWatchlistById(@Param('id') id: string, @Req() request: Request) {
+    const user = request['user'];
+
+    if (!user.sub) {
+      throw new UnauthorizedException('User not found');
+    }
+
     return await this.watchlistService.getWatchlistById(Number(id));
   }
 
-  @Delete(':userId/:contentId')
+  @Delete('/:Id')
   @ApiOperation({ summary: 'Remove content from watchlist' })
   @ApiResponse({
     status: 200,
@@ -109,12 +135,20 @@ export class WatchlistController {
   })
   @ApiResponse({ status: 404, description: 'Watchlist entry not found.' })
   async delete(
-    @Param('userId') userId: string,
-    @Param('contentId') contentId: string,
+    @Param('id') id: string,
+    @Req() request: Request,
   ): Promise<string> {
+    const user = request['user'];
+
+    if (!user || user.sub) {
+      throw new UnauthorizedException(
+        'User not authorized to delete this watchlist entry',
+      );
+    }
+
     return await this.watchlistService.deleteWatchListEntry(
-      Number(userId),
-      Number(contentId),
+      Number(user.sub),
+      Number(id),
     );
   }
 }
